Update 40f6af0346a99aa1cd1d555a4e9cce62c7f9634603ee406615833dc8c8d00367.pem
Removed a space byte after the certificate footer.

The reason for this tiny modification comes from an OpenJDK bug; Java X509Factory throws a CertificateException when parsing certificates with a whitespace after the header or footer. While this bug has been patched in OpenJDK 12, it is still present in the previous versions:
https://bugs.openjdk.java.net/browse/JDK-8208602

This is the only certificate causing this exception.